Fix migration script unique index create for GPDB6
In Greenplum Database, unique indexes are allowed only if the columns of
the index key are the same as (or a superset of) the Greenplum
distribution key. However, on a 5X cluster, when creating a unique index
on an empty table that does not satisfy this rule, the database will not
error out, but instead implicitly change the distribution key to satisfy
this rule. This disabled on 6X and is no longer the behavior for this
edge case; An error will be thrown. Create the unique index on the
distribution key to allow test case to work for 5X and 6X.

GPDB reference commit:
https://github.com/greenplum-db/gpdb/commit/2505f1f5998f69c6a9ce526277022eefe5e1a976