fix unique and primary key constraints initialize data migration script
The data migration scripts create alter statements to drop the
constraints from both the root and child partitions in that order.

Since dropping unique constraints cascade to the child partitions when
the alter child partition statement is executed it fails since the
constraint was already removed when the root partition alter statement
was run. However, the error message is misleading saying "ERROR: can't
drop a constraint from "table_with_unique_constraint_p_1_prt_1"; it is
part of a partitioned table." Rather than something like constraint does
not exist.

This problem only occurs for unique constraints which cascade since
primary key constraints do not cascade to their child partitions.

Thus, as a workaround reverse the order in which we drop constraints
from the child partitions up to the root partitions.