Add gp_fastsequence test to verify duplicate ctid fix
It was discovered that not upgrading gp_fastsequence resulted in broken
indexes in upgraded clusters. This is because the ctids generated from
gp_fastsequence values may be duplicated upon new inserts in AO and AOCO
tables. Index implementation is esstentially a map where ctids acts like
keys and the table's rows are values. This test is to make sure indexes
are still functional after upgrade.

gpdb 6X reference commit:
https://github.com/greenplum-db/gpdb/commit/c90639e6e4081411122b450b4433da96dc6f416a

Co-authored-by: Jimmy Yih <jyih@vmware.com>