revert: simplify expected mirror failure check
Due to a GPDB 5X issue upgrading the primaries results in an invalid
checkpoint upon starting. The checkpoint needs to be replicated to the
mirrors with rsync or gprecoverseg. When upgrading the mirrors during
finalize the checkpoint is replicated. In copy mode the 5X source
cluster mirrors do not start causing gpstart to return a non-zero exit
status. Ignore such failures, as gprecoverseg is executed to bring up
the mirrors. Running gprecoverseg is expected to not take long.

Co-authored-by: Kalen Krempely <kkrempely@vmware.com>