Use cdw instead of mdw in CI scripts and yaml templates
Our CI uses a custom tool called CCP to provision GPDB test clusters
on the cloud for our automated tests. There was a recent change that
changed the hostnames mdw and smdw (master and standby master) to cdw
and scdw (coordinator and standby coordinator) as part of the
inclusive terminology changes. The master data directory was also
changed to use coordinator instead of master. Technically we could pin
our CCP to the v2.6.0 tag but updating the hostnames here as well is
better so that we keep ourselves aligned with the latest CCP code. The
CI scripts and yaml templates have been updated accordingly.