Do not VACUUM FREEZE all tables in partition heap table tests
Running VACUUM FREEZE on all tables during the partition heap tables
tests were creating extra unexpected entries when running the
gp_fastsequence tests. This is also a race condition waiting to happen
since this tests is run in parallel with many other tests. It also has
the potential to affect xmin testing. Changed to only vacuum the
subpartition that is being tested.

Co-authored-by: Jimmy Yih <jyih@vmware.com>