ci: fail early if patching the target dump file failed or was skipped
When the ci filter patch is out-of-date, the patching command may skip
when it finds a diff in the patch that it cannot apply. We were not
checking the return code for the patch command so the job would
continue and run the long diff against the source and target dumps
which may eventually fail depending on if there were any GPDB changes
to the dump output. When looking at the dump diffs as the result of
the patch apply skip, the unexpected diffs could be deceptive when the
ci filter patch is what needs to be updated. Check the return code and
fail early if the patching fails or is skipped.

As a result of checking the return code, it was also found that the
retail_demo.diff filter patch was no longer needed because the `START
WITH ` filter was integrated into scripts/filters/rules6x.go a while
back. Just in case we need it later, keep the empty file around for
now.