Add acceptance test for list partitioned tables with custom type
A partition table can be list partitioned by a custom type but this
previously had issues when running through upgrade because the CREATE
TABLE assumes the custom type's equality operator is in the pg_catalog
schema or in search_path. Since the SQL dump from pg_dump explicitly
sets the search_path to blank, only pg_catalog was searched which
would result in the pg_restore section of upgrade to fail. This has
been fixed in a recent GPDB commit and we should be able to support
this scenario now.

GPDB commit reference:
https://github.com/greenplum-db/gpdb/commit/d18208cf0bb82e2ba78121547f44f9615196744e