Fix revert after early initialize exit
With the migration scripts being generated and applied at the
beginning of the initialize command now, starting up the hub and
creating the config file is done much later. Both are required for
revert to succeed but we allow users to quit early during the
migration scripts part of Initialize. If the user quits Initialize
during that early phase, running revert will fail. To fix the issue,
we create and prepopulate the config file and start the hub prior to
the migration script logic in Initialize. New step logic has been
introduced as well to allow us to skip over unnecessary revert
substeps.

Co-authored-by: Jimmy Yih <jyih@vmware.com>