refactor ArchiveLogDirectories
The current implementation does not work if a substep after archive log
directory fails and revert or finalize are re-run then the logs won’t be
properly archived.

Derive the archive log directory name each time it is called. This is
useful for substep idempotence if revert or finalize need to be re-run
then the archive log directory will succeed with the same upgradeID, but
different timestamp.