refactor hub responses
Use the type system to our advantage and introduce protobuf types for
Cluster and Segment. Ideally, we would use these directly in the code to
avoid needing to serialize and deserialize between our Cluster and
SegConfig structs, but it seems protobufs do not support maps where the
key is a struct or complex type.

Future work could also change Segconfig's PrimaryRole and MirrorRole to
use protobuf's enum type similar to how ClusterDestination is currently
used.