alwaysRun certain substeps
When recently debugging a live customer issue when upgrading the
coordinator failed resulted in needing to always run certain substeps
rather than skipping them. The source cluster was started for inspection
and then gpupgrade execute was re-run. However, stopping the source
cluster substep was skipped since it was previously run causing
pg_upgrade to fail since it expects the cluster to be stopped. Thus,
always run some substeps which are needed if a failure occurs
downstream.