Add StepStore interface
Also rename NewStepStore to NewStepFileStore to indicate it is a
specific implementation of the StepStore interface. By introducing an
interface we can now create a mock step store for easy testing. In the
future it also easily allows us to switch from a file based store to
sqlite or similar, and continues our alignment with step/step.go
framework.