add --pg-upgrade-skip-checks
Since pg_upgrade checks can take a long time for large clusters add
--pg-upgrade-skip-checks to speed up internal testing feedback cycles.

This is supported for both initialize and execute as hidden flags. The
command is "not" persisted in config.json. Thus, when specifying in
initialize it is not persisted to execute and will need to be manually
specified on the command line each time.