Check that master and standby master are synchronized on GPDB 5X as well
It's very important to make sure the master and standby master are
synchronized before running Initialize and Execute. This is because
Revert will rsync the standby master data directory onto the master
data directory which can cause severe master corruption if the standby
master not up-to-date. We have this check for GPDB 6X+ but it was
seemingly missed for 5X because the catalog view gp_stat_replication
does not exist on 5X. However, we're only checking the master and
standby master replication status so there's no need to use the
cluster-wide gp_stat_replication view; we can simply use
pg_stat_replication which will only report the replication status of
the individual segment (in this case, just the master segment).