move integration tests into test directory
Consolidate all (non-unit) test files into the test directory.

The motivation is that there are soon to be golang acceptance tests to slowly
replace the BATs tests. As we were working on where to best place the upcoming
golang acceptance tests the organization of the existing integration tests were
discovered.

Co-authored-by: Kevin Yeap <kyeap@vmware.com>