CI: run gpcheckcat after e2e tests
For extra validation post-upgrade for end-to-end tests run gpcheckcat.
This is in addition to validating the standby and mirrors post-upgrade
by failing over and failing back.

Note, we do not run gpcheckcat or validate mirrors after the pg_upgrade
acceptance tests or multi-host gpupgrade acceptance tests. This would
require leaving the cluster intact after running the tests which is
cumbersome during dev iteration and upon failures.