serialize and deserialize cluster using gob
We need a way to take the complex internal golang structure of a cluster
and send it from the hub to the cli process and have all those same
helper functions (ie: RunGreenplumCmd(), CheckActiveConnections(), etc.)
be available on that structure. One way of doing this is serializing to
bytes using gob, sending it via protobuf, and deserializing it back.

A better solution would be to convert current cluster object to use a
protobuf which we then extend to add on our helper functions. That way
there is no need to serialize and deserialize when sending over
protobuf. Its all just protobuf both internally and externally when
sending.

However, that is a fairly big change especially around the Tablespace
object. This is because protobuf maps only support keys which are scalar
types (ie: ints, strings, etc.) and not other objects. Also protobuf
maps can’t have a value which is another map. We violate these things
for the Tablespace object requiring a fairly sizeable refactor.