allow CLI substreams to properly write to streams
The previous BufferedStreams wasn't fully allowing CLI substeps to write
to stdout and sterr and would only output in verbose mode. Swich to
step.OutStreams. This is useful when running RunGreenplumCmd() which
sets the commands stdout and stderr to the passed in streams. This
matches the desired expectation found in the Hub.