ensure log prefix is printed to hub and agent logs
Currently, the hub and agent substeps write to both stdout/stderr “and”
the log file. Similar functionality should likely be done for the CLI as
well. That is, writing to stdout will also write to the log.

The log prefix contains the timestamp which is critical for debugging
and support. Previously the log file handle was passed to
MultiplexedStream which was not correctly adding the log prefix. Ratherm
simply calling log.Print() does print the prefix.

Thus, refactor multiplexedStream to call log.Print to print out the
prefix with timestamp. Doing so removes the need for Close() and its
associated functions. When printing to the log trim leading and trailing
whitespace to avoid awkward multi-line whitespace in the log file.