fix flakey intialize aceptance test
This test is flakey since the internal code ranges over a map to
determine overlapped ports. Golang does not preserve ordering when
ranging over maps. Thus, have the test just check for a number rather
than a specific number for the port. This is easier than changing the
production code.