enable 6 > 7 and 7 > 7 ci jobs
The purpose of this is to enable the 6 > 7 and the 7 > 7 jobs without
breaking the current 5 > 6 jobs. Following commits to start making
tweaks so the 6 > 7 pg_upgrade tests are able to run. We expect them to
fail, but we need them to be in a state where ci at least runs them.