pg_upgrade nonupgradable test detecting views with removed functions
Views that use removed functions will cause upgrade to fail. This
happens during metadata restore on the target cluster because pg_restore
will error trying to create a view using functions that do not exist
anymore. This is not ideal as we could be many hours into upgrade before
a view using a removed function causes pg_upgrade to fail. This check
calls a support function on the source cluster to check if views using
removed functions exist.